Extract Bucket, MappedBucket and Service from __init__
